# Batcher buffers merging optimization