Batcher consists of multiple parts:
1. Collecting information on batch groups.
2. Figuring out individual batches based on mesh differences.
3. Merging meshes buffers with additional transformation.

This PR focuses on the third case by decomposing some math functions that are called frequently, which results in significant performance gains in some of our tests.

In a scene with 749 batches, on test machine currently it takes:
~1658ms to apply position transformations and 612ms on normals transformations.

With this PR these numbers become:
~216ms to apply position transformations and 155ms on normals transformations.

In this specific case we have **x7.7 time speed up** on positions and **x3.9 times speed up** on normals.
We've tested multiple cases, and gains can vary from x1.5 to x10, depending on number of meshes per batch and how large these meshes are.

I confirm I have read the [contributing guidelines](https://github.com/playcanvas/engine/blob/master/.github/CONTRIBUTING.md) and signed the [Contributor License Agreement](https://docs.google.com/a/playcanvas.com/forms/d/1Ih69zQfJG-QDLIEpHr6CsaAs6fPORNOVnMv5nuo0cjk/viewform).