@mvaligursky 2024-11-25
cherry picked to engine v2 as well